requirements.txt
dj-database-url==0.5.0
Django==2.0.7
django-extensions==2.1.0
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
